Start using new content format
Blog posts are now defined by a text file that's a mix of
CommonMark and key-value sections, separated by --- (sort of).

For now, the key-value sections can only be used to define images.
Those images are then marked up differently in the now slightly
prettier template with CSS.

This still needs a lot of work.