Fix confusing value of "dir" in secrets.json
Only write the random directory name containing the post to the
secrets.json config file, not any other part of the path. This
means secrets.json contains no information about where a post
was published, only what the structure of the blog is.

Move creation of the directory out of Blog.write() and into
Post.write(), since it is never guaranteed to exist, even if
the post was published before.