Zoom in to panorama if range is smaller than view
If the vertical range on a cropped panorama photo is smaller than
the default viewer zoom level, zoom in more. Avoid showing black
areas.

This obsoletes my change to minFov. That was another attempt at
solving the same problem. But the only thing that did was change
the default zoom level. That's halfway between minFov and maxFov.