Avoid fractional zoom level in cropped panoramas
I'm zooming in to cropped panoramas slightly if their latitude
range does not fill the view. That way I avoid seeing black
around the edges. But my earlier fix wasn't enough.

I presume it wasn't enough due to rounding somewhere. Floor the
zoom level to a whole number to restrict the range a bit more.

I'm not sure if this is really the right kind of fix in the right
place, but most other places would be in the Photo Sphere Viewer
or in three.js outside of my control.