Refactor content format and post implementation
Instead of a post being created from a single text file with mixed
sections of CommonMark and metadata that links to media files,
each post is now created from a directory. The directory can
contain CommonMark files and images. A CommonMark file with the
same name as an image (minus extension) becomes a comment on the
image. The post is formed by the alphabetically ordered content.

Several other parts of the implementation also changed:
* Encryption moves into a Cipher class.
* JSON config encoding and decoding moves into json.JSONDecoder
  and json.JSONEncoder subclasses.
* A blog now consists of Blog, Post, PostSection and PostComponent
  classes. Support for more content file types can be added by
  registering additional PostComponent subclasses.
* Blog posts no longer have a title. This and the other changes
  means Jinja2 is no longer needed and the requirement is removed.